+++ assets/js/dashboard/util/filters.js
@@ -23,16 +23,25 @@ export const NO_CONTAINS_OPERATOR = new Set(['goal', 'screen'].concat(FILTER_MOD
 export const EVENT_PROPS_PREFIX = "props:"
 
 export const FILTER_OPERATIONS = {
+  is: 'is',
   isNot: 'is_not',
   contains: 'contains',
-  is: 'is'
+  does_not_contain: 'does_not_contain'
 };
 
-export const OPERATION_PREFIX = {
+export const FILTER_OPERATIONS_DISPLAY_NAMES = {
+  [FILTER_OPERATIONS.is]: 'is',
+  [FILTER_OPERATIONS.isNot]: 'is not',
+  [FILTER_OPERATIONS.contains]: 'contains',
+  [FILTER_OPERATIONS.does_not_contain]: 'does not contain'
+}
+
+const OPERATION_PREFIX = {
   [FILTER_OPERATIONS.isNot]: '!',
   [FILTER_OPERATIONS.contains]: '~',
   [FILTER_OPERATIONS.is]: ''
 };
+
 
 export function supportsIsNot(filterName) {
   return !['goal', 'prop_key'].includes(filterName)
@@ -52,17 +61,6 @@ try {
 }
 
 const ESCAPED_PIPE = '\\|'
-
-function escapeFilterValue(value) {
-  return value.replaceAll(NON_ESCAPED_PIPE_REGEX, ESCAPED_PIPE)
-}
-
-function toFilterQuery(type, clauses) {
-  const prefix = OPERATION_PREFIX[type];
-  const result = clauses.map(clause => escapeFilterValue(clause.toString().trim())).join('|')
-  return prefix + result;
-}
-
 
 export function getLabel(labels, filterKey, value) {
   if (['country', 'region', 'city'].includes(filterKey)) {
@@ -141,19 +139,19 @@ export function cleanLabels(filters, labels, mergedFilterKey, mergedLabels) {
   return result
 }
 
+const EVENT_FILTER_KEYS = new Set(["name", "page", "goal", "hostname"])
 
-// :TODO: New schema for filters in the BE
 export function serializeApiFilters(filters) {
-  const cleaned = {}
-  filters.forEach(([operation, filterKey, clauses]) => {
-    if (filterKey.startsWith(EVENT_PROPS_PREFIX)) {
-      cleaned.props ||= {}
-      cleaned.props[getPropertyKeyFromFilterKey(filterKey)] = toFilterQuery(operation, clauses)
-    } else {
-      cleaned[filterKey] = toFilterQuery(operation, clauses)
+  const apiFilters = filters.map(([operation, filterKey, clauses]) => {
+    let apiFilterKey = `visit:${filterKey}`
+    if (filterKey.startsWith(EVENT_PROPS_PREFIX) || EVENT_FILTER_KEYS.has(filterKey)) {
+      apiFilterKey = `event:${filterKey}`
     }
+    clauses = clauses.map((value) => value.toString())
+    return [operation, apiFilterKey, clauses]
   })
-  return JSON.stringify(cleaned)
+
+  return JSON.stringify(apiFilters)
 }
 
 export function fetchSuggestions(apiPath, query, input, additionalFilter) {